requirements.txt
appdirs==1.4.3
astroid==1.5.2
decorator==4.0.11
Django==1.11
djangorestframework==3.6.2
ipython==6.0.0
ipython-genutils==0.2.0
isort==4.2.5
jedi==0.10.2
lazy-object-proxy==1.3.1
mccabe==0.6.1
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.14
ptyprocess==0.5.1
Pygments==2.2.0
pylint==1.7.1
pyparsing==2.2.0
pytz==2017.2
requests==2.14.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
wrapt==1.10.10
